+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [v0.3.3] - 2017-01-09
+
 ### Added
 
 - Support for building a custom sysroot when compiling natively.
@@ -213,7 +215,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Initial release
 
-[Unreleased]: https://github.com/japaric/xargo/compare/v0.3.2...HEAD
+[Unreleased]: https://github.com/japaric/xargo/compare/v0.3.3...HEAD
+[v0.3.2]: https://github.com/japaric/xargo/compare/v0.3.2...v0.3.3
 [v0.3.2]: https://github.com/japaric/xargo/compare/v0.3.1...v0.3.2
 [v0.3.1]: https://github.com/japaric/xargo/compare/v0.3.0...v0.3.1
 [v0.3.0]: https://github.com/japaric/xargo/compare/v0.2.3...v0.3.0